Updated tools from Project tools to Global tools.
Tools are now build targeting .NET Core 2.1, although they can still be used to deploy projects using other versions of .NET Core but .NET Core 2.1 is required for running these tools.
Since the tools can now be run outside of a project directory. Add a check for commands that are going to build to make sure the are being run in directories that contain a .NET project.