Use MSBuild to Determine Target Framework for a Project
There are more ways to specify the target framework for a project than
setting it directly in the `TargetFramework` element in a
project. There are other unambiguous scenarios which this tool should
be able to support:

1. TargetFramework in the project file
1. TargetFrameworks in the project file, with one framework specified
1. Either 1 or 2 explicitly imported via `<Project Import=""/>`
1. Either 1 or 2 implicitly imported via a means like
   Directory.Build.props

We should be able to support anything that MSBuild does by deferring
the determination of target framework to MSBuild. For cases in which
it returns multiple results (or no results, I suppose, though I don't
know how that would build in the first place), the same fallback
mechanism is triggered.